services/planner-dapr-agent/requirements.txt
# OpenAI Agents SDK with Dapr integration
openai-agents[dapr]>=0.7.0

# Dapr SDK for state/pub-sub operations
dapr>=1.14.0

# Dapr workflow extension for true workflow durability
dapr-ext-workflow>=1.14.0

# FastAPI for HTTP endpoints
fastapi>=0.100.0
uvicorn>=0.22.0

# Jinja2 for HTML templates
jinja2>=3.1.0
aiofiles>=23.0.0

# Pydantic for data validation
pydantic>=2.0.0

# HTTP client for cross-namespace event forwarding
httpx>=0.24.0

# Environment variables support
python-dotenv>=1.0.0

# OpenInference instrumentation for Phoenix/Arize LLM observability
# Auto-instruments OpenAI SDK calls to capture generation spans
openinference-instrumentation-openai>=0.1.0
openinference-semantic-conventions>=0.1.0

# OpenTelemetry for distributed tracing
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-httpx>=0.41b0
opentelemetry-instrumentation-logging>=0.41b0

# Kubernetes Agent Sandbox client for isolated execution
agentic-sandbox @ git+https://github.com/kubernetes-sigs/agent-sandbox.git@main#subdirectory=clients/python/agentic-sandbox-client
